+++ admin/src/main/resources/io/buoyant/admin/js/src/admin.js
@@ -2,10 +2,9 @@
 
 define([
   'jQuery',
-  'Handlebars',
-  'text!template/router_option.template',
+  'template/compiled_templates',
   'bootstrap'
-], function($, Handlebars, templateRsp) {
+], function($, templates) {
   function initialize(removeRoutersAllOption) {
     // highlight current page in navbar
     var path = window.location.pathname;
@@ -18,18 +17,17 @@ define([
     }
 
     return $.get("config.json").done(function(routersRsp) {
-      var template = Handlebars.compile(templateRsp);
+      var template = templates.router_option;
       var routers = routersRsp.routers;
 
-      //Not every page supports a mult-router view!
-      if(!removeRoutersAllOption) {
-        routers.unshift({label: "all"});
-      } else {
-        if (!matches && routers.length > 0) {
-          selectRouter(routers[0].label || routers[0].protocol);
-        }
+      if (!matches && routers.length > 0) {
+        selectRouter(routers[0].label || routers[0].protocol);
       }
 
+      //Not every page supports a mult-router view!
+      if(!removeRoutersAllOption || routers.length === 0) {
+        routers.push({label: "all"});
+      }
       var routerHtml = routers.map(template);
 
       $(".dropdown-menu").html(routerHtml.join(""));
@@ -43,11 +41,6 @@ define([
   }
 
   function selectRouter(label) {
-    if (label === "all") {
-      unselectRouter();
-      return;
-    }
-
     var uriComponent = "router=" + encodeURIComponent(label);
     var re = /router=(.+)(?:&|$)/
     if (window.location.search == "") {