+++ admin/src/main/resources/io/buoyant/admin/js/src/router_client.js
@@ -8,15 +8,14 @@ define([
   'src/query',
   'src/success_rate_graph',
   'src/bar_chart',
-  'text!template/metric.partial.template',
-  'text!template/router_client.template'
+  'template/compiled_templates'
 ], function($, _, Handlebars,
   Utils,
   Query,
   SuccessRateGraph,
   BarChart,
-  metricPartialTemplate,
-  routerClientTemplate) {
+  templates
+) {
 
   var LoadBalancerBarChart = function($lbContainer) {
     function getColor(percent) {
@@ -50,7 +49,7 @@ define([
   }
 
   var RouterClient = (function() {
-    var template = Handlebars.compile(routerClientTemplate);
+    var template = templates.router_client;
 
     var metricToColorShade = {
       "max": "light",
@@ -83,9 +82,8 @@ define([
       });
     }
 
-    function renderMetrics($container, client, summaryData, latencyData, clientColor) {
+    function renderMetrics($container, client, summaryData, latencyData) {
       var clientHtml = template({
-        clientColor: clientColor,
         client: client.label,
         latencies: latencyData,
         data: summaryData
@@ -136,24 +134,26 @@ define([
     }
 
     return function (metricsCollector, routers, client, $container, routerName, colors, shouldExpandInitially) {
-      var metricPartial = Handlebars.compile(metricPartialTemplate);
+      var metricPartial = templates["metric.partial"];
       Handlebars.registerPartial('metricPartial', metricPartial);
 
       var $contentContainer = $container.find(".client-content-container");
+
+      var $headerLine = $container.find(".header-line");
+      var colorBorder = "2px solid" + colors.color;
 
       var $metricsEl = $container.find(".metrics-container");
       var $chartEl = $container.find(".chart-container");
       var $toggleLinks = $container.find(".client-toggle");
       var $lbBarChart = $container.find(".lb-bar-chart");
 
-      var clientColor = colors.color;
       var latencyLegend = createLatencyLegend(colors.colorFamily);
       var metricDefinitions = getMetricDefinitions(routerName, client.label);
 
       var $expandLink = $toggleLinks.find(".client-expand");
       var $collapseLink = $toggleLinks.find(".client-collapse");
 
-      renderMetrics($metricsEl, client, [], [], clientColor);
+      renderMetrics($metricsEl, client, [], []);
       var chart = SuccessRateGraph($chartEl.find(".client-success-rate"), colors.color);
       var lbBarChart = new LoadBalancerBarChart($lbBarChart);
 
@@ -169,8 +169,14 @@ define([
 
       function toggleClientDisplay(expand) {
         if (expand) {
+          $contentContainer.css({"border": colorBorder});
+          $headerLine.css("border-bottom", "0px");
+
           metricsCollector.registerListener(metricsHandler, getDesiredMetrics);
         } else {
+          $contentContainer.css({'border': null});
+          $headerLine.css({'border-bottom': colorBorder});
+
           metricsCollector.deregisterListener(metricsHandler);
         }
 
@@ -187,7 +193,7 @@ define([
         chart.updateMetrics(getSuccessRate(summaryData));
         lbBarChart.update(summaryData);
 
-        renderMetrics($metricsEl, client, summaryData, latencies, clientColor);
+        renderMetrics($metricsEl, client, summaryData, latencies);
       }
 
       function getDesiredMetrics(metrics) {