+++ admin/src/main/resources/io/buoyant/admin/js/src/router_clients.js
@@ -1,16 +1,18 @@
 "use strict";
 
 define([
-  'jQuery', 'Handlebars',
+  'jQuery',
   'src/router_client',
   'src/combined_client_graph',
-  'text!template/router_client_container.template'
-], function($, Handlebars,
+  'template/compiled_templates'
+], function($,
   RouterClient,
   CombinedClientGraph,
-  routerClientContainerTemplate) {
+  templates
+) {
   var RouterClients = (function() {
     var EXPAND_CLIENT_THRESHOLD = 6;
+    var TRANSFORMER_RE = /(\/%\/[^\$#]*)?(\/[\$#]\/.*)/;
 
     function assignColorsToClients(colors, clients) {
       return _.reduce(clients, function(clientMapping, client, idx) {
@@ -25,7 +27,7 @@ define([
     }
 
     return function (metricsCollector, routers, $clientEl, $combinedClientGraphEl, routerName, colors) {
-      var clientContainerTemplate = Handlebars.compile(routerClientContainerTemplate);
+      var clientContainerTemplate = templates.router_client_container;
 
       var clients = routers.clients(routerName);
       var colorList = colors;
@@ -47,10 +49,23 @@ define([
       function initializeClient(client) {
         $clientEl.show();
         var colorsForClient = clientToColor[client.label];
+        var match = ('/' + client.label).match(TRANSFORMER_RE);
         var $container = $(clientContainerTemplate({
           clientColor: colorsForClient.color,
-          client: client.label
+          prefix: match[1],
+          client: match[2]
         })).appendTo($clientEl);
+        if (match[1]) {
+          var $transformerPrefix = $container.find(".transformer-prefix");
+          var $clientSuffix = $container.find(".client-suffix");
+
+          $container.find(".client-id")
+            .click(function() {
+              $transformerPrefix.toggle("slow", function() {
+                $clientSuffix.toggleClass("is-first", !$transformerPrefix.is(":visible"));
+              });
+            });
+        }
 
         return RouterClient(metricsCollector, routers, client, $container, routerName, colorsForClient, expandClients);
       }