+++ linkerd/admin/src/main/scala/io/buoyant/linkerd/admin/AdminHandler.scala
@@ -70,10 +70,10 @@ class AdminHandler(navItems: Seq[NavItem]) extends HtmlView {
         <head>
           <title>linkerd admin</title>
           <link type="text/css" href="files/css/lib/bootstrap.min.css" rel="stylesheet"/>
+          <link type="text/css" href="files/css/fonts.css" rel="stylesheet"/>
           <link type="text/css" href="files/css/dashboard.css" rel="stylesheet"/>
           <link type="text/css" href="files/css/logger.css" rel="stylesheet"/>
           <link rel="shortcut icon" href="files/images/favicon.png" />
-          <link href='https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,300,600' rel='stylesheet' type='text/css'>
           $cssesHtml
         </head>
         <body>