+++ linkerd/protocol/http/src/integration/scala/io/buoyant/linkerd/protocol/TlsTerminationTest.scala
@@ -17,7 +17,7 @@ class TlsTerminationTest extends FunSuite with Awaits {
           s"""
              |routers:
              |- protocol: http
-             |  baseDtab: |
+             |  dtab: |
              |    /p/dog => /$$/inet/127.1/${dog.port} ;
              |    /s/clifford => /p/dog ;
              |  servers:
@@ -56,7 +56,7 @@ class TlsTerminationTest extends FunSuite with Awaits {
           s"""
              |routers:
              |- protocol: http
-             |  baseDtab: |
+             |  dtab: |
              |    /p/dog => /$$/inet/127.1/${dog.port} ;
              |    /s/clifford => /p/dog ;
              |  servers: